+++ src/Bridge/Nette/NettePsrContainer.php
@@ -5,7 +5,11 @@ declare(strict_types=1);
 namespace Multitron\Bridge\Nette;
 
 use Nette\DI\Container;
+use Nette\DI\MissingServiceException;
+use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
+use Psr\Container\NotFoundExceptionInterface;
+use RuntimeException;
 
 class NettePsrContainer implements ContainerInterface
 {
@@ -15,10 +19,12 @@ class NettePsrContainer implements ContainerInterface
 
     public function get($id): mixed
     {
-        if (!class_exists($id)) {
-            return $this->container->getByName($id);
+        try {
+            return $this->container->getByType($id, false) ?? $this->container->getByName($id);
+        } catch (MissingServiceException $e) {
+            throw new class($e->getMessage(), 0, $e) extends RuntimeException implements NotFoundExceptionInterface {
+            };
         }
-        return $this->container->getByType($id);
     }
 
     public function has($id): bool